Enhance the customization abilities
This introduces a format and hence we can drop the mappings for mode symbols as
people can configure this as they want. There is also no need to specify the
concatenation differently since we now use a format string and can place things
in any way we want.

This also adds some nice default hover style to the mode line button, as well as
binding the scroll-wheel up/down to change the tab-width.

The max tab-width size, min tab-width size as well as the steps for each
increment/decrement is now configurable.