Convert between tabs/spaces
This also overhauls the menu definition as it was getting out of hand
when I was adding some extra things.

I also don't see the need to customize the menu like before, so I have
removed that. If there is interest to be able to customize the mouse
actions on the mode-line which I usually don't see other modes do, we
could just enable customization of the bindings and not the whole menu.

Remove the forced reloading of the mode-line as it seems it reloads fine
by itself.

Clean up the mode-line format stuff in general.